Create player character, Set game as main scene
I created a player character that is a triangle pointed to the right. I
set the motion mode to floating as this is a top-down 2D game. I gave
the player a collision shape that is slightly smaller than the polygon
for forgiveness in game. I set custom input for movement in the project
settings as well. I added the default CharacterBody2D script,
removed unnecessary parts, and added deceleration to smooth out the
movement. I chose not to add acceleration to the player because it made
the game feel less responsive.

I also added the player to the game scene and set the game scene as the
main scene for testing.