Implement shooting mechanic for player
I wanted the player to be able to shoot a simple bullet, so I created a
HandleShooting function on the player. I then set the global position of
the bullet and add it as a child to the player's parent. When I added
the bullet as a child of the player, I was unintentionally able to
control the bullet after firing it. With the level set as the bullet's
parent, I can no longer control the bullet after shooting.

I also velocity calculations on the bullet to make it move across the
screen.

Finally, I added a VisibleOnScreenNotifier to the bullet. Then, when the
bullet is no longer visible on screen, I delete it from the scene tree.
Before I did this, the scene tree was getting flooded with insane
amounts of bullets. This would have inevitably bogged down the game
after a period of playing.

I still need to delete the bullets from the scene tree when they hit a
wall, but I will save that for another commit.