Create an enemy that moves linearly to the left
I created an enemy script that allows the enemy to move to the left at a
speed set within the editor.
I noticed that overriding PhysicsProcess isn't as smooth as overriding
Process, but I don't want to use Process for movement since that's not
recommended. I will have to mess around with it later and attempt to
polish it. It looks choppy right now, but I want it to be as smooth as
the players frame rate.