Create an enemy spawner, Fix namespaces
I created an EnemySpawner node that chooses a random point within the
area and instantiates an enemy at that position. The EnemySpawner class
doesn't track timers as that will be managed by the Game class. All this
does is choose where to spawn the enemy and instantiate it at the
correct position.

I also changed some of the namespaces to follow C# namespace naming
conventions. Originally, I had Player and Enemy as namespaces and
classes which was causing issues. I made the namespaces plural to avoid
the errors.