Create skeleton for player pickups
Originally, I was going to differentiate between the player and gained
troops, but I decided against that. I renamed the troop_spawner scene to
player_pickup because it makes more sense in this context. I created a
script for the player_pickup to be added to later. I also created a
Marker2D in the Game scene to indicate where player_pickups should spawn
initially.