Implement pickups that spawn and move in Game
I created a player_pickup scene that handles the movement of the pickup.
I created a timer in the Game scene and connected it to a function that
spawns a player pickup every 2 seconds. I need to move this to a
PickupSpawner class as it is crowding the Game class. I also need to
change the name of the collision layer the player_pickup is on since I
no longer refer to added players as troops.