Create a player spawner, Organize Game class
I created a player spawner class that only contains one function at the
moment. The function accepts a Vector2 parameter and will spawn the
player at that position.

I removed the player as a node in the Game scene and replaced it with
the player spawner. Then, in the _Ready function of the Game class, I
spawn the first player at it's origin. I reorganized and renamed some of
the variables in the Game class since they didn't all follow the same
naming patterns. Finally, I created a GameStart function in the Game
class for when I implement failing and restarting.