Create PlayerManager
I created a PlayerManager class that will be able to manage everything
player related as I add it. I changed the Game class to reference
PlayerManager instead of PlayerSpawner directly. I also changed how the
Player adds bullets to the scene because they were being added as
children of the PlayerManager node after my changes. That is not correct
since the PlayerManager only manages the Player, not Bullets as well.