Create SignalBus autoload
I needed a signal for the PlayerPickup that is accessible globally to
reduce coupling between PlayerManager and PlayerPickups. Creating a
SignalBus autoload (based on the EventBus pattern) seemed like the best
solution.

Since signals must be local to the class they're called from in C#, I
created a public Instance property in the SignalBus class. Then,
whenever I need to emit a signal, I do it from SignalBus.Instance.
Similarly, when I subscribe to the event, I subscribe to the
SignalBus.Instance event.