Spawn Player below Player that hit PlayerPickup
When a Player collides with the PlayerPickup block, the PlayerManager
handles spawning a new Player. I tried spawning the new Player directly
on the Player that collided with the PlayerPickup, but it caused a lot
physics issues. I solved this by offsetting the spawn point by 5 pixels
to the right. This also prevents the new Player from immediately
touching a wall which will eventually kill the player.