Spawn Player below Player that hit PlayerPickup
When a Player collides with the PlayerPickup block, the PlayerManager
handles spawning a new Player. I tried spawning the new Player directly
on the Player that collided with the PlayerPickup, but it caused a lot
physics issues. I solved this by offsetting the spawn point by 5 pixels
to the right. This also prevents the new Player from immediately
touching a wall which will eventually kill the player.

It's worth noting that spawning all of these players as individual nodes
and having them each shoot a bullet every physics tick is extremely
taxing on the system. When more than 10 players exist and I position
them at the back wall and shoot, the game lags very hard. I will most
likely have to optimize this, but I want to implement a slower fire rate
before exploring advanced optimization.