Kill player when they collide with wall
I created a new SignalBus signal that can be emitted when the player
collides with a wall. I created it in the SignalBus class instead of the
Wall class because I needed the PlayerManager to receive the signal.
That would have required much stricter coupling without the use of the
SignalBus.

I created a new public function on the Player that frees it from the
queue. This is for the PlayerManager to call.

I created a public function on the PlayerManager that can kill a
specific Player. I set this function as the observer for the
PlayerTouchedWall signal.

Currently, the way I track the amount of players that are left is
extremely inaccurate. I have not troubleshot why yet, but I will have to
before I can accurately tell when the game is lost.