Only allow a PlayerPickup to be used once
There was a weird issue where multiple Players could collide with the
same PlayerPickup in the same physics tick and spawn multiple Players
from a single PlayerPickup. I tried disabling collision of the
PlayerPickup on the first collision, but that has to be deferred. This
causes the collision disable to occur too late, resulting in the same
issue.

I solved this by creating an private field _isCollected and setting it
to true on the first collision. I check if it is true before handling
the collision. This results in the PlayerPickup being single use which
is the intended functionality.