Only allow track Player death once
There was a weird issue where a Player could alert the PlayerManager of
it's death multiple times resulting in frequent negative _playerCount in
PlayerManager. I created an public property IsDead on the Player that
gets set to true when the Player collides with a wall the first time. I
then check if isDead is true before handling death logic. This prevents
multiple collisions from a single Player counting as multiple deaths.