Create UIManager autoload
This may not be the best solution to this problem, but I created an
autoload that will handle many of the UI elements. I expect to use this
for the bigger menus like the main menu, pause menu, upgrades screen,
etc. Although enemy health bars would be UI, I don't know if it makes
sense to put that in this autoload. I'll figure that out when/if it's an
issue.