Add PauseMenu to UIManager on Game scene ready
I created a pause menu that can be activated when the Game scene is
ready. The pause menu does not have much functionality yet, but it can
toggle pause within the Game scene.

Currently, the UIManager holds a public reference to the instantiated
PauseMenu node. I'm not sure that this is the best way to handle this,
but it works for now. I'll use it like this until it doesn't work.