Fix weird pause menu bug
When the resume button and ui_cancel key were pressed at the same time,
it caused a weird issue with the pause menu. It would stay paused but
only had the resume button. It seems the resume button and quit button
may have been overlaid, but I'm not sure. Changing resume button's
function to only unpause fixed the issue.