Show game over menu when player count reaches zero
Since the PlayerManager handles everything player related, I made it
emit a signal from the SignalBus whenever a Player dies. Then, in the
Game script, it checks if the playerCount is zero. If it is, the game
over scene is added as a child of the game.

I added the game over scene as a child of the game since that scene
should only be shown in the game. If I delete the game node from the
scene tree, the game over scene will also be deleted. I will change the
pause menu to do this as well.