QueueFree enemies and pickups that exit screen
Originally, I was not removing objects from the scene tree that weren't
handled during the game, i.e. the enemy getting killed or pickup getting
picked up. I added some simple code and a VisibleOnScreenNotifier2D to
the enemy and player_pickup scenes to handle this.

The reason for implementing this is because I like when the game objects
continue spawning on the game over screen. If I didn't remove these
nodes after they were off screen, the game would gradually lag until it
eventually crashed on the game over screen. I avoid that issue by doing
this.