No pause on game over, No game over during game
There was an issue where the pause menu could still be activated even
when the game over screen was up. I didn't like the overlaid UIs, so I
disabled the pause menu when the game has ended. Likewise, I disabled
the game over menu while the game is still live.