Restart game when Play Again button pressed
The SignalBus emits a signal when the restart button is pressed in the
game over screen. The Game class listens to this signal and runs the
GameStart() function.

Currently, this function does not restart the game; It just spawns
another player at the start. All Enemies and PlayerPickups that were in
the level still are. I will change this eventually.