Refactor UIManager to SceneManager
I didn't like that the pause menu and game over menu were in the
autoload even though they were only relevant to the Game scene. I
realized that this autoload would be better implemented as a scene
manager that switches between main scenes, not in-game scenes. The
SceneManager class is pretty bare now, but it will have more once I
implmement the main menu.