Connect main menu and game scene through autoloads
I created the SceneManager autoload to be able to switch between scenes
like the main menu and game more easily. It kind of works at the moment,
but there are definitely some issues.

Currently, it seems the starting values are not reset properly, which I
will fix in a future commit.