Fix disposed object errors, rename private fields
Whenever I restarted the game, I would get disposed object errors upon
touching PlayerPickup or dying. This was because the freed PlayerManager
was still marked as an observer of some global signals. To fix this, I
unsubscribe from those signals when deleting the node from the tree.

I also had some private fields for the menus that were not following C#
naming conventions, so I fixed that as well.

Finally, there was a GameStarted signal in the SignalBus that wasn't
being emitted from anywhere. I deleted this and stopped listening to it
in the Game class.