Implement a shooter queue
Originally, shooting was handled within the Player object, but that did
not allow for Players to shoot at different times. Now, the
PlayerManager holds a list of all the Players that are currently in the
scene tree as well as who the current Player is. When the player wants
to shoot, only the current Player's shoot function is called.

A shared fire rate has not been implemented yet, so regardless of how
many Players there are, a bullet is shot once every physics tick. Having
more Players just causes the individual fire rate to go down since they
have to wait more physics ticks before they are the next Player allowed
to shoot.

There is a bug with the current index that I discussed in a comment. In
short, I don't think it will be an issue once I implement a slower fire
rate, but it does occur quite often when the fire rate is once per
physics tick.