Track time survived
There was no scoring system previously, so the player just tried to
survive as long as possible for no reason. I have implemented a scoring
system that tracks how long a player has survived. Currently, it only
outputs it to the console, but I will display this value on the game
over screen in a future commit.