Reorganize, Refactor
I had a one of the timeout functions separated from the other timer
functions, so I reorganized so they are together.

I also needed to call signal connections and disconnections from other
functions besides the _Ready and _ExitTree function, so I moved them to
their own functions. I then call them in the GameStart and GameOver
functions. I do this because, when the game ends, the Game class
continued processing signals even though the game was over. This is
unintended functionality, so I disconnect from all signals when the
GameOver function is called