Display time survived on game over screen
To give the player more of a reward for surviving longer, I wanted to
display how long the player survived after they lose. This feels better
compared to the game ending and nothing shows.

I achieve this functionality through emitting a signal that includes
time survived when the game ends. This is currently only handled by the
game over menu, but it could later be handled by other systems (such as
an achievement system).