Janky fix to bullet collision
I was having issues with bullets going through walls and enemies
occasionally. I managed to eliminate this by stretching the bullet
collision shape further to the right. You would think this would cause
issues with the bullet colliding too early, but since the bullet moves
so quickly, it's not noticeable.