Fix shooting bug after losing a lot of Players
I discovered an issue that occurred where I could not shoot anymore
after quickly losing a lot of Players. I believe this happened because
the current player index was slightly outdated and it was trying to call
a dead Player to shoot.

I fixed this by updating the current player index immediately before
calling a Player's shoot function. This seems like it could still rarely
cause the bug, but I have not been able to get it to occur. I will
revisit this if I ever encounter the bug again.