Track enemy health in code
Originally, I was killing enemies as soon as they were hit with a
bullet. I want enemy health to increase as the player survives longer,
so that implementation no longer works. I now make the enemy take damage
when it collides with a bullet, and each time the enemy takes damage, it
checks if it should free itself if it's health is zero or lower.

I thought I was going to have an issue with bullets colliding multiple
times with the same enemy, but I was incorrect. The OnAreaEntered signal
only gets handled once when the bullet first overlaps the enemy. I was
thinking it got called every physics tick while they still overlapped,
but that was wrong. No bugs to fix today.