Fix bullets not despawning when off screen
If a Player was all the way against the right edge of the screen and
shot, the bullet would actually spawn outside of the viewport. This
would cause the VisibleOnScreenNotifier2D to never signal that the
bullet went off screen because it was always off screen. I fixed this by
making the VisibleOnScreenNotifier2D trail behind the bullet a bit which
makes it start on screen regardless of where the Player shoot from.