Show enemy health bar once damaged
I already had a way to track enemy health in code, but the player could
not tell what the enemy's health was even once damaged. I created a
health bar for each enemy that only shows once it is damaged for the
first time. This keeps the screen less cluttered with health bars
because it can be assumed that a newly spawned enemy is full health.