Call Enemy damage function from Bullet
Originally, I was checking if a Bullet and Enemy were colliding in two
different classes, Bullet and Enemy. Bullet's class would just delete
the Bullet, and Enemy's class would handle the damage. This doesn't make
much sense though because why is the Enemy handling a Bullet's damage?

Instead, I made it so the Bullet calls the Enemy's damage function
directly. This feels better from the previous solution because now the
Bullet functionality is contained in the Bullet class.