Create IDamageable interface that enemy implements
I noticed that Bullets had to check if the object it collided with was
an enemy and saw how this could cause coupling issues in the future.
If I create a different enemy type, I don't want to have to change the
bullet to get the new enemy working. Therefore, I created an IDamageable
interface that damageable entities can implement.

I only added this interface to the enemy for now, but it needs to be
added to the Player after I get the HealthComponent working.