Create health component that Enemy uses
I started having a lot of entities with health, so I decided to make a
health component they can all use. This is better than creating a base
class the entities inherit from because it is more flexible. I did not
change any functionality of the enemy; I just moved it's health code to
the health component.

I still need to add the health component to the Player class as well.