Implement IDamageable and health on Player
In an effort to make the game scale better, I created an IDamageable
interface and health component that can be used by whichever classes
need them. The Player class finds these very useful, so I implemented
them to the best of my knowledge. I had to change some of the other
classes (Enemy and Wall) because they were calling a signal that was no
longer necessary. That helps reduce some of the global signal bus usage.