Abstract movement strategy out of Enemy class
I want different enemies to be able to move differently, so I abstracted
the simple movement pattern out of the Enemy class. I set up a strategy
pattern that can be used for future movement styles I want to implement.
This pattern should make it very simple to plug in new movement styles
on new enemies.