Improve enemy movement abstraction
I abstracted enemy movement out of the Enemy class in a previous commit,
but it wasn't type safe at compile time. This meant that an incompatible
resource could be attached to an EnemyResource, and it would still
compile. When the game tried to spawn an Enemy, it would error because
the EnemyResource was not set in the editor properly.

To add compile time safety, I gave the BaseMovementStrategy and
StandardMovementStrategy classes the GlobalClass attribute. This allows
the editor to see these classes and know that only concrete classes that
inherit BaseMovementStrategy are valid options.

I also moved the EnemyResource file to entities/enemies because it made
more sense there instead of globals.