Start creating Enemy with squiggly movement
I created most of what I need for the new Enemy type. The only thing
that still needs completed is actually making the movement follow a sine
wave. Currently, the enemy travels Southwest for about half the screen
and then stabilizes into a sine wave, but I want it to stay as the same
sine wave the whole time. I'll figure that out later.