Creating an enemy that has squiggly movement
This new movement strategy allows an enemy to move in a sine wave
pattern, making it harder for the player to hit the enemy.

Currently, there is a bug where enemies can go through walls. I think I
need to change the wall's type so enemies can collided with it. Enemies
cannot currently collide with walls because they are both Area2Ds.