Change wall type to StaticBody2D
I had the wall as an Area2D, but this was not allowing the enemy to
collide with it, which I need. The first step to fixing this was
changing the wall's type to a StaticBody2D. To keep the collision
functionality the same, I added an Area2D as a child with it's own
collision shape. I then connected the appropriate signals to maintain
the same functionality.