Change enemy type to CharacterBody2D
When the enemy type was Area2D, it could not collide with anything in
the world. This is not the behavior I want, so I had to either create my
own collision code or change the enemy type to something that can
collide by default. The latter was definitely easier.

I added an Area2D for the enemy collision which feels weird, but it
probably makes sense in the long run.

I had to change the movement strategies to accept CharacterBody2Ds
instead of CollisionObject2Ds because I wanted to use the MoveAndSlide
function from the CharacterBody2D class. This made updating the movement
very easy.