Create 10 hard-coded waves
Originally, I only had 2 simple waves of standard enemies. Now, I have
10 combinations that get more difficult over time. It feels somewhat
unbalanced at the moment because the final waves are very difficult. I
might increase the time between fast enemy spawns to fix this, but I'm
not sure yet. Another alternative is to make the fast enemy slower, but
that might defeat the purpose of the fast enemy.

While I do have endless mode programmed, it only repeats the final,
hard-coded wave. This is not the intended final functionality, but I
have not created the endless wave logic yet.