Create procedural wave generation
Previously, when the user beat the final hardcoded wave, that last
difficult wave would be repeated until the player failed or it became so
easy that it got boring. I created a WaveBuilder class that procedurally
generates a different wave every time it's called. This will reduce the
repetition a player experiences after the hardcoded waves have been
beaten.

This code is currently very ugly. It is functional at this point, but I
NEED to go back and make it more readable.