Clean up WaveBuilder code
I went back through the WaveBuilder code and fixed the comments that
were incorrect or unnecessary. I planned on breaking the code up into
multiple functions, but I don't think that's required. It is a large
function, but it's not something I plan on changing or working on
frequently, so I left it as-is.