Create a HUD that displays current wave
I can tell which wave I'm on without a HUD because I made the waves, but
the player did not have any visual cue that they were making progress. I
created a HUD that displays the current wave at the top of the screen. I
may change this to make it pop in when the wave changes and fade slowly
so it isn't always on the screen, but that idea might be scrapped.