Display final wave on game over screen
The player was only told how long they survived on the game over screen.
I want them to also see the highest wave they reached, so I added it to
one of the labels on the game over screen. Currently, the player can
still see what wave they reached on the hud, but if change that to not
always be visible, this ensures the player still sees that info.