Fix error when enemies to spawn is zero
When the random ratio generated for a specific enemy type was zero, it
was causing an error later when creating the enemy sequence. This is
because the random amount of enemies to spawn would try and choose
between 1 and 0. This caused a reversed min max error on the random
number generator and would not return a wave.

Instead of writing a bunch of code to check if each enemy had enemies to
spawn initially, I just changed the ratio range from 0-100 to 1-100.
When using the ceiling function later, this ensures that each enemy type
will be spawned at least once per wave.

While this isn't the most robust solution, it was the easiest to
implement for a function I don't intend to touch again.