Implement running animation
When the player is not moving, they should show the idle animation.
Similarly, when they are moving, they should show a running animation.
As long as the change in position is greater than a specified minimum,
the player is considered to have moved. This means that a player that is
actually moving shows the running animation while a player that is
trying to move while blocked by something will not show the running
animation.

Note that there is a jitter issue with the players trying to get as
close as possible. I believe this is caused by the ellipse collision
shape, but I will look for a solution.