Implement shooting animation
Previously, when the player shot, there would be no visual indication
from the player that they shot, i.e. recoil. Now, the gun recoils when
a player shoots which looks nice.

Since I have the gun tracks in the same animation as the body tracks,
things can look weird sometimes. For example, if the player is currently
near the end of the run animation and shoots, the player will appear to
float for a few frames while shooting since the code continue the run
animation while playing the shoot animation. This could be fixed by
separating the player and the gun, but I didn't want to do that for
such a small project.