Implement player death animation
Previously, the PlayerManager was in charge of deleting the player when
they died. However, this would cause the player death animation to be
cut short. I moved despawn functionality to the player and played the
death animation when the player died.

I also had to account for the player still being interactable after
death. This was an easy enough fix by stopping processing and physics
processing and disabling collision.