Implement bullet sprite, Update bullet spawn pos
Previously, the bullet was just a white rectangle which did not match
the color palette of the asset pack I'm using. I switch the Polygon2D
out for a Sprite2D that uses a bullet from the asset pack. I did some
small adjustments to make the bullet easier to see in game.

I also moved the spawn position of the bullet because it was spawning in
the player's chest. I shifted it right to make it spawn at the end of
the gun barrel. This is not a robust solution for different weapons, but
that's fine since I'm not adding more weapons. YAGNI.