Remove player death on wall collision
Originally, I thought it might be fun to have the walls of the viewport
be lava that needed to be avoided. However, having to get close enough
to the wall to kill an enemy and far enough away to not die was
annoying. It was not a fun game mechanic, so I commented it out. I
didn't fully remove it in case I go back to the wall script to add
something, but I will not be adding that feature back in.