Set up Y sorting
Enemies would commonly be drawn incorrectly and have their legs in front
of another enemy. I easily fixed this by making the Game Node a Node2D
and enabling Y sorting on that as well as some other affected nodes. The
gun can be drawn in front of the walls which is a little weird, but it's
good enough for now.