Play gun sound when player shoots
I implemented a gunshot sound effect with the AudioManger singleton.
This singleton will not scale well with tons of sounds effects, but it
works for my purpose here.

I chose the maximum limit for the gunshot sound effect because there
will be many Players shooting at once. It would be weird to play fewer
sound effects than shots occurring.