Create unique sound effect for gunshot
Previously, I just created the SoundEffect for the gunshot within the
AudioManager inspector, but I didn't see a way to rename the
SoundEffects I created. That was going to get unwieldy very quickly, so
I created the gunshot SoundEffect elsewhere and then referenced it from
AudioManager's inspector. This allows me to see what each SoundEffect is
for by just checking the name.