Add restart and main menu buttons to pause menu
Previously, the player could only resume or quit the game from the pause
menu, but there was no way to restart or return to the main menu. That
felt weird, so I copied code from the game over scene to handle that.

If I were to do this correctly, I would create utility functions that
handle restarting and returning to the main menu as this will occur in
many different places throughout the codebase. However, it's not super
necessary for this prototype.