Refactor wave complete audio trigger
Previously, WaveManager was in charge of when the wave complete audio
played. However, this caused issues with the enemy spawning on the main
menu. Since WaveManager controlled the audio trigger, the wave complete
audio was playing on the main menu even though waves shouldn't matter. I
moved the audio handling to the Game class since that audio should only
play during the game.