Create an animated background for the main menu
Previously, the main menu was an ugly, gray screen that didn't show the
game feel at all. I reused some game logic in a new class and created an
animated background for the main menu. This background just plays the
waves that will appear in game on a different level.