Parse JSON into List of Sections
Before I can auto-create the credits in the credit scene, I need to
parse the credit data from the JSON file. I created a couple classes to
define objects parsed from the JSON and parsed it within the Credits
class. The data is not viewable by the player yet, but the data is
stored in a list.