Abstract JsonParser from CreditsMenu
I will be using the JsonParser elsewhere in my code, so I abstracted it
out of the CreditsMenu class.

I also did that so I could more easily attach the script directly to the
node that will be appended. This allows me to not have to reference a
specific node in the parent CanvasLayer.