Change Labels to RichTextLabels in JsonParser
When using Labels for the credits content, I was unable to select text.
Since links are unclickable at the moment, it made it very inconvenient
to navigate to a link from the credits. I changed the Labels to
RichTextLabels so that text could be selected, copied, and pasted by the
user.