Hide main menu when credit button is pressed
Previously, when I opened the credits menu, the main menu buttons showed
through since the credits menu is slightly transparent. I did not like
that, so I created a function that toggles the menu content visibility.
It hides itself when I click the credits button, and it toggles itself
again when I close the credits menu. This functionality is abstract
enough that I can easily use it with other menus I need to create.