Abstract custom JSON deserializer
Previously, I had a JsonParser class that was doing everything related
to JSON for the CreditMenu class. I want to generalize this so I can use
it for my tutorial page as well.

I abstracted the custom JSON deserializer from JsonParser so that it can
be reused in other menus with different JSON structures.