Create new tutorial menu scene
This scene is almost exactly the same as credits_menu spare a few name
changes. I could have generalized the credits_menu to be used as a
pop-up menu, but that's not necessary for this small of a project. If I
keep working on this game, I will refactor this at some point.