Finalize Tutorial menu
Previously, the tutorial data was loaded but never displayed. Now, all
of the tutorial data is displayed appropriately. I was going to use a
VBoxContainer to display the tutorials, but I felt that a Tree is much
cleaner. It provides collapsible menus for each section that makes the
entire menu look cleaner, even though I'm still only using default
assets.