Free Player when it exits screen, SSBCIDFLDITCM
Previously, when the Player was not controlled and it exited the screen,
it would not be freed. That could cause a memory leak over time, so I
fixed that.

I also set the running animation to play when the Player is not
controlled and set _isControlled to false by default. I still need to
add a public method that allows setting the Player's controlled state.