Refactor PlayerSpawner, Change Pickup to Player
Originally, having the PlayerManager control Player spawns worked, but
since I wanted to change PlayerPickups to uncontrolled Players, I needed
the PlayerSpawner somewhere else. Game made the most sense to be the
parent of PlayerSpawner, so that's what I did. I modified all the logic
in Player and PlayerManager to be able to handle the PlayerSpawner as a
child of Game.

I also removed all instances of PlayerPickups and replaced them with
Players with a control flag. This looks much more polished compared to a
big, purple block floating across the top lane. It also prevents players
from getting a Player stuck up top for easy Player recruitment (although
I may add this as a feature if I continue working on this).