Fix fullscreen on Windows, Display controls
I was misinterpreting the different window modes because they are
labelled weirdly. I was using fullscreen because I thought that was the
borderless fullscreen option. I was not using exclusive fullscreen
because I thought that was the old fullscreen method where the screen
flashed black until the window was fullscreen.

While fullscreen mode worked fine on Linux (as most things do), it did
not work as expected on Windows (as most things don't). It left a 1px
border around the edge of the window that did not allow the viewport to
properly fullscreen. I switched to using exclusive fullscreen, and this
issue is now gone. It also bypasses the compositor on Linux which is an
added bonus.

I also added a label on the main menu instructing the user how they can
toggle fullscreen.